+++ ios/Classes/SocialSharePlugin.m
@@ -17,22 +17,27 @@
 
 - (void)handleMethodCall:(FlutterMethodCall*)call result:(FlutterResult)result {
     if([@"shareInstagramStory" isEqualToString:call.method]){
+        //Sharing story on instagram
       NSString *stickerImage = call.arguments[@"stickerImage"];
       NSString *backgroundTopColor = call.arguments[@"backgroundTopColor"];
       NSString *backgroundBottomColor = call.arguments[@"backgroundBottomColor"];
       NSString *attributionURL = call.arguments[@"attributionURL"];
         NSString *backgroundImage = call.arguments[@"backgroundImage"];
-
+        //getting image from file
       NSFileManager *fileManager = [NSFileManager defaultManager];
       BOOL isFileExist = [fileManager fileExistsAtPath: stickerImage];
       UIImage *imgShare;
       if (isFileExist) {
+          //if image exists
           imgShare = [[UIImage alloc] initWithContentsOfFile:stickerImage];
       }
+        //url Scheme for instagram story
       NSURL *urlScheme = [NSURL URLWithString:@"instagram-stories://share"];
-        
+        //adding data to send to instagram story
        if ([[UIApplication sharedApplication] canOpenURL:urlScheme]) {
+           //if instagram is installed and the url can be opened
            if ( [ backgroundImage  length] == 0 ){
+              //If you dont have a background image
              // Assign background image asset and attribution link URL to pasteboard
              NSArray *pasteboardItems = @[@{@"com.instagram.sharedSticker.stickerImage" : imgShare,
                                             @"com.instagram.sharedSticker.backgroundTopColor" : backgroundTopColor,
@@ -45,12 +50,14 @@
              [[UIPasteboard generalPasteboard] setItems:pasteboardItems options:pasteboardOptions];
                  
                [[UIApplication sharedApplication] openURL:urlScheme options:@{} completionHandler:nil];
+                 //if success
                  result(@"sharing");
            } else {
                result(@"this only supports iOS 10+");
            }
            
        }else{
+           //if you have a background image
            NSFileManager *fileManager = [NSFileManager defaultManager];
            BOOL isFileExist = [fileManager fileExistsAtPath: backgroundImage];
            UIImage *imgBackgroundShare;
@@ -116,73 +123,121 @@
   }else if([@"copyToClipboard" isEqualToString:call.method]){
       NSString *content = call.arguments[@"content"];
       UIPasteboard *pasteboard = [UIPasteboard generalPasteboard];
+      //assigning content to pasteboard
       pasteboard.string = content;
       result([NSNumber numberWithBool:YES]);
   }else if([@"shareTwitter" isEqualToString:call.method]){
 //      NSString *assetImage = call.arguments[@"assetImage"];
       NSString *captionText = call.arguments[@"captionText"];
-      
+      NSURL *urlScheme = [NSURL URLWithString:@"twitter://"];
+      if ([[UIApplication sharedApplication] canOpenURL:urlScheme]) {
+          //check if twitter app exists
       NSString *urlSchemeTwitter = [NSString stringWithFormat:@"twitter://post?message=%@",captionText];
-      NSURL *urlScheme = [NSURL URLWithString:urlSchemeTwitter];
+      NSURL *urlSchemeSend = [NSURL URLWithString:urlSchemeTwitter];
       if (@available(iOS 10.0, *)) {
-          [[UIApplication sharedApplication] openURL:urlScheme options:@{} completionHandler:nil];
+          [[UIApplication sharedApplication] openURL:urlSchemeSend options:@{} completionHandler:nil];
           result(@"sharing");
       } else {
           result(@"this only supports iOS 10+");
       }
+          
+          }else{
+                     result(@"cannot find Twitter app");
 
+                 }
   }else if([@"shareSms" isEqualToString:call.method]){
         NSString *msg = call.arguments[@"message"];
         NSString *urlstring = call.arguments[@"urlLink"];
+      NSURL *urlScheme = [NSURL URLWithString:@"sms://"];
+
         NSString* urlTextEscaped = [urlstring stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding];
         NSURL *url = [NSURL URLWithString: urlTextEscaped];
+      //check if it contains a link
         if ( [ [url absoluteString]  length] == 0 ){
+            //if it doesn't contains a link
         NSString *urlSchemeSms = [NSString stringWithFormat:@"sms:?&body=%@",msg];
         NSURL *urlScheme = [NSURL URLWithString:urlSchemeSms];
+            if ([[UIApplication sharedApplication] canOpenURL:urlScheme]) {
         if (@available(iOS 10.0, *)) {
             [[UIApplication sharedApplication] openURL:urlScheme options:@{} completionHandler:nil];
             result(@"sharing");
         } else {
             result(@"this only supports iOS 10+");
         }
+                }else{
+                           result(@"cannot find Sms app");
+
+                       }
     }else{
+        //if it does contains a link
         NSString *urlSchemeSms = [NSString stringWithFormat:@"sms:?&body=%@",msg];
         NSString *urlWithLink = [urlSchemeSms stringByAppendingString:[url absoluteString]];
-               NSURL *urlScheme = [NSURL URLWithString:urlWithLink];
+               NSURL *urlSchemeMsg = [NSURL URLWithString:urlWithLink];
+        if ([[UIApplication sharedApplication] canOpenURL:urlScheme]) {
                if (@available(iOS 10.0, *)) {
-                   [[UIApplication sharedApplication] openURL:urlScheme options:@{} completionHandler:nil];
+                   [[UIApplication sharedApplication] openURL:urlSchemeMsg options:@{} completionHandler:nil];
                    result(@"sharing");
                } else {
                    result(@"this only supports iOS 10+");
                }
+        }else{
+            result(@"cannot find Sms app");
+
+        }
     }
       }else if([@"shareSlack" isEqualToString:call.method]){
       //      NSString *content = call.arguments[@"content"];
            
             result([NSNumber numberWithBool:YES]);
         }else if([@"shareWhatsapp" isEqualToString:call.method]){
-        //      NSString *content = call.arguments[@"content"];
-             
+              NSString *content = call.arguments[@"content"];
+             NSString * urlWhats = [NSString stringWithFormat:@"whatsapp://send?text=%@",content];
+             NSURL * whatsappURL = [NSURL URLWithString:[urlWhats stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding]];
+             if ([[UIApplication sharedApplication] canOpenURL: whatsappURL]) {
+                 [[UIApplication sharedApplication] openURL: whatsappURL];
+                 result(@"sharing");
+             } else {
+                 result(@"cannot open whatsapp");
+             }
+
               result([NSNumber numberWithBool:YES]);
           }else if([@"shareTelegram" isEqualToString:call.method]){
-          //      NSString *content = call.arguments[@"content"];
-               
+          NSString *content = call.arguments[@"content"];
+                     NSString * urlScheme = [NSString stringWithFormat:@"tel://msg?text=%@",content];
+                     NSURL * telegramURL = [NSURL URLWithString:[urlScheme stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding]];
+                     if ([[UIApplication sharedApplication] canOpenURL: telegramURL]) {
+                         [[UIApplication sharedApplication] openURL: telegramURL];
+                         result(@"sharing");
+                     } else {
+                         result(@"cannot open Telegram");
+                     }
                 result([NSNumber numberWithBool:YES]);
             }
           else if([@"shareOptions" isEqualToString:call.method]){
             NSString *content = call.arguments[@"content"];
               NSString *image = call.arguments[@"image"];
-              NSFileManager *fileManager = [NSFileManager defaultManager];
-                        BOOL isFileExist = [fileManager fileExistsAtPath: image];
-                        UIImage *imgShare;
-                        if (isFileExist) {
-                            imgShare = [[UIImage alloc] initWithContentsOfFile:image];
-                        }
-               NSArray *objectsToShare = @[content, imgShare];
-               UIActivityViewController *activityVC = [[UIActivityViewController alloc] initWithActivityItems:objectsToShare applicationActivities:nil];
-              UIViewController *controller =[UIApplication sharedApplication].keyWindow.rootViewController;
-               [controller presentViewController:activityVC animated:YES completion:nil];
-                result([NSNumber numberWithBool:YES]);
+              //checking if it contains image file
+              if ([image isEqual:[NSNull null]] || [ image  length] == 0 ){
+                  //when image is not included
+                  NSArray *objectsToShare = @[content];
+                           UIActivityViewController *activityVC = [[UIActivityViewController alloc] initWithActivityItems:objectsToShare applicationActivities:nil];
+                          UIViewController *controller =[UIApplication sharedApplication].keyWindow.rootViewController;
+                           [controller presentViewController:activityVC animated:YES completion:nil];
+                            result([NSNumber numberWithBool:YES]);
+              }else{
+                  //when image file is included
+                  NSFileManager *fileManager = [NSFileManager defaultManager];
+                                         BOOL isFileExist = [fileManager fileExistsAtPath: image];
+                                         UIImage *imgShare;
+                                         if (isFileExist) {
+                                             imgShare = [[UIImage alloc] initWithContentsOfFile:image];
+                                         }
+                                NSArray *objectsToShare = @[content, imgShare];
+                                UIActivityViewController *activityVC = [[UIActivityViewController alloc] initWithActivityItems:objectsToShare applicationActivities:nil];
+                               UIViewController *controller =[UIApplication sharedApplication].keyWindow.rootViewController;
+                                [controller presentViewController:activityVC animated:YES completion:nil];
+                                 result([NSNumber numberWithBool:YES]);
+              }
             }else {
     result(FlutterMethodNotImplemented);
   }